ai-engine/requirements.txt
flask
flask-cors
python-dotenv
google-genai
pdfplumber
spacy
serpapi
pandas
statsmodels
textblob
gunicorn